% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_opened.R
\name{file_opened}
\alias{file_opened}
\title{file_opened}
\usage{
file_opened(path)
}
\arguments{
\item{path}{A path to the file that has to be checked}
}
\value{
boolean value, TRUE if file is open, FALSE if closed
}
\description{
Checks if a file is opened (helper function for plot2ppt)
}
