% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not in}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{First R object of any kind.}

\item{y}{Second R object of any kind.}
}
\value{
Logical.
}
\description{
Not in
}
\examples{
x <- 1:10
y <- 11:20
x \%!in\% y

}
\references{
\url{https://stackoverflow.com/a/5831829}
}
