% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnk_against_random.R
\name{knnk_against_random}
\alias{knnk_against_random}
\title{Compare average degree of the neighbors-KNN against simulated networks}
\usage{
knnk_against_random(x, n.sim, n.cor)
}
\arguments{
\item{x}{igraph newtork.}

\item{n.sim}{integer. Number of simulated networks to create.}

\item{n.cor}{integer. Number of processing cores.}
}
\value{
Data frame whose first column is the degree range (\eqn{k_{min}, \dots, k_{max}}) of x, the second column is the observed knnk and the rest are knnks of random simulations.
}
\description{
Calculates average degree of the neighbors-KNN of a given network and then perform \code{n.sims} degree-preserving simulations of the input network and calculates its corresponsing KNN. This is useful to check the validity of a network, as if it is different from the random expectation, then there must be a biologically plausible reason behind generating the observed network KNN
}
