% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_alpha_on_deleteion.R
\name{my_alpha_on_deletion}
\alias{my_alpha_on_deletion}
\title{Calculate alpha after targeted deletion of nodes}
\usage{
my_alpha_on_deletion(network, n.cor)
}
\arguments{
\item{n.cor}{}
}
\description{
Calculate alpha after targeted deletion of nodes
}
