% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_plotCounts.R
\name{my_plotCounts}
\alias{my_plotCounts}
\title{Custom per gene plotCounts}
\usage{
my_plotCounts(ensembl, dds.object, plot.title = NULL,
  pvalue.contrasts = NULL, pdf.name = NULL, ...)
}
\arguments{
\item{ensembl}{Character. One or many gene identifiers in the same format (i.e ensembl ID, entrez ID, symbol, etc) as in the \code{rownames(assay(dds.object))}.}

\item{dds.object}{DESeqDataSet. Constructed using \code{DESeq()}.}

\item{plot.title}{Character. Vector of the same length and order as \code{ensembl}, specifying the tile name for the plot (e.g gene name). If a vector is supplied and it does not match with \code{ensembl}, then a \code{NA} will be assigned to the plot tile. Default prints no plot title.}

\item{pvalue.contrasts}{Character. Each contrast must be of the form \code{c("tretament_1", "treatment_2")} and many contrasts can be especified by creating a list of contrasts. See \link[ggpubr]{stat_compare_means}.}

\item{...}{other arguments to pass to \link[ggpubr]{stat_compare_means}.}
}
\value{
A list of ggplot2 plots.
}
\description{
Creates a customised ggplot2 plot of counts per gene from a \code{DESeqDataSet} object. It also add the pvalue significance to a user specified list of contrasts with \link[ggpubr]{stat_compare_means}.
}
\examples{
\dontrun{
genes <- c("ENSG00000026025", "ENSG00000198836")
gene.names <- c("vim", "opa1")
my.contrasts <- list(c("tibolona", "vehiculo"), c("tibolona", "tib_pal"), c("tibolona", "palmitato"), c("tibolona", "DMEM"))
aesir <- my_plotCounts(genes, dds, gene.names, my.contrasts)
aesir
}

\dontrun{
# Supliying additional arguments to stat_compare_means()
genes <- c("ENSG00000026025", "ENSG00000198836")
gene.names <- c("vim", "opa1")
my.contrasts <- list(c("tibolona", "vehiculo"), c("tibolona", "tib_pal"), c("tibolona", "palmitato"), c("tibolona", "DMEM"))
aesir <- my_plotCounts(genes, dds, gene.names, my.contrasts, method = "t.test")
aesir
}

}
\seealso{
\code{\link[DESeq2]{plotCounts}}, \code{\link[ggpubr]{stat_compare_means}}, \code{\link[ggsignif]{geom_signif}}
}
