% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_targeted_robustness.R
\name{my_targeted_robustness}
\alias{my_targeted_robustness}
\title{Targeted network robustness}
\usage{
my_targeted_robustness(network, n.cor)
}
\arguments{
\item{n.cor}{}
}
\description{
Targeted network robustness
}
