% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_up_down.R
\name{my_up_down}
\alias{my_up_down}
\title{Extract number of up and down regulated genes per contrast}
\usage{
my_up_down(dds, contrasts)
}
\arguments{
\item{dds}{DESeqDataSet object. Can be constructed with the \code{DESeq} function.}

\item{contrasts}{list. Each element must be a valid contrast name or contrast formula as expected by \code{DESeq2::results()}.}
}
\value{
Data frame with 5 columns: 1. "up" the number of upregulated genes in the specified contrast, 2. "down" the number of downregulated genes in the specified contrast, 3. the percentage of upregulated genes out of the total DE genes, 4. the percentage of downregulated genes out of the total DE genes and 5. the contrast name.
}
\description{
Creates a useful summary dataframe contrast wise to compare thenumber of upregulated and downregulated genes in a bulk RNAseq experiment analysed with DESEq2. This data frame can then be easily ploted to get insight about the contrast with more changes.
}
\examples{

}
