% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_vulcano.R
\name{my_vulcano}
\alias{my_vulcano}
\title{Vulcano plot for each contrast in DESeq2}
\usage{
my_vulcano(dds_path, contrast_name, my_lfcThreshold = 0,
  my_alpha = 0.1, output_pdf_path, ...)
}
\arguments{
\item{dds_path}{DESeqDataSet. Obtained from running \link[DESeq2]{DESeq}. The object itself can be suplied or a quoted string defining the path to the \code{dds.rds} object (e.g. \code{my/folder/dds.rds}).}

\item{contrast_name}{list of \code{contrast} as described in \code{\link[DESeq2:results]{results}}. Level names can be seen by calling \code{resultsNames(dds)}.}

\item{my_lfcThreshold}{integer. Decision rule cutoff (e.g. DESeq2 is 0.10).The smaller the more restrictive the threshold will be and fewer genes will pop up.}

\item{my_alpha}{integer. Cutoff for the Log2 Fold Change (e.g. DESeq2 default is 0, however 1.5 is also very common). The bigger the more restrictive the threshold will be and fewer genes will pop up.}

\item{...}{Additional arguments to the ploting function \link[EnhancedVolcano]{EnhancedVolcano}. A coomon one is to define a hard lim for the x axis by \code{xlim = c(-5, 5)}.}
}
\value{
Either a list of ggplot \code{EnhancedVolcano} plots or a message indicating the path to the plots. In the former, there will be a \code{ggplot} object per contrast.
}
\description{
After doing the differential expression step, the next step is to prioritise genes to validate or focus on. This function enables the quick ploting of those genes inusing vulcano plots. Additionally, the arguments \code{alpha} and \code{lfcThreshold}for the function \link[DESeq2]{results} can be tunned to be more liberal or restrictive when choosing interesting DE genes.
}
\examples{
\dontrun{
dds <- makeExampleDESeqDataSet()
dds <- DESeq(dds)

df <- as.data.frame(colData(dds))
}
data(dds)
contrast_name <- list(c("condition","palmitato", "vehiculo"), c("condition","tib_pal", "vehiculo"), c("condition", "tibolona", "tib_pal"), c("condition","tibolona", "vehiculo"), c("condition", "palmitato", "tibolona"))

 vulcano_per_contrast <- my_vulcano(
 dds_path = dds,
 contrast_name = contrast_name,
 my_alpha = 0.05,
 my_lfcThreshold = 1
)


}
