% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots2ppt.R
\name{plots2ppt}
\alias{plots2ppt}
\title{plots2ppt}
\usage{
plots2ppt(plot, doc, path)
}
\arguments{
\item{plot}{a ggplot object}

\item{doc}{a rpptx object from the package officer, doc <- officer::read_pptx()}

\item{path}{A path to write the file}
}
\value{
a pptx file
}
\description{
takes a ggplot object and writes it into a power point presentation
}
\examples{
\dontrun{
library(ggplot2)
data(mpg)
carplot <- ggplot(data=mpg) + geom_boxplot(aes(x=manufacturer, y=displ))
doc <- officer::read_pptx()
plots2ppt(carplot, doc, path="my_presentation.pptx")
}
}
