% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmake_pairwise_plot.R
\name{bbmake_pairwise_plot}
\alias{bbmake_pairwise_plot}
\title{Create pairwise comparison table with Cohen's d and CIs}
\usage{
bbmake_pairwise_plot(
  emm,
  pw_table,
  formula,
  group1,
  group2,
  aesthetic = aes(x = xvar, y = yvar),
  y.adjust = 0
)
}
\arguments{
\item{emm}{An emmGrid object (output from emmeans())}

\item{pw_table}{A table_df object (output from bbmake_pairwise_table())}

\item{formula}{Formula object used to create emm}

\item{group1}{"String" First group for significance markers for pairwise comparisons}

\item{group2}{"String" Second group for significance markers for pairwise comparisons}

\item{aesthetic}{Aesthetic used for the plot. Default = aes(x = xvar, y = yvar)}

\item{y.adjust}{Float or list of floats representing the y adjustment for each significance marker}
}
\value{
A ggplot2 plot object
}
\description{
Create pairwise comparison table with Cohen's d and CIs
}
