% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmake_pairwise_table.R
\name{bbmake_pairwise_table}
\alias{bbmake_pairwise_table}
\title{Create pairwise comparison table with Cohen's d and CIs}
\usage{
bbmake_pairwise_table(pw)
}
\arguments{
\item{pw}{An emmGrid object (output from pairs() from emmeans)}
}
\value{
A data.frame object
}
\description{
Create pairwise comparison table with Cohen's d and CIs
}
\examples{
pw_table <- bbmake_pairwise_table(pw)
}
