% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots2ppt.R
\name{plots2ppt}
\alias{plots2ppt}
\title{plots2ppt}
\usage{
plots2ppt(plot, doc, path)
}
\arguments{
\item{plot}{a ggplot object}

\item{doc}{a rpptx object from the package officer}

\item{path}{A path to write the file}
}
\value{
a pptx file
}
\description{
takes a ggplot object and writes it into a power point presentation
}
